Added functionality to MatrixMapGenerator
Added second map object. This object is a Map<String,int[][]>. A layer of the map is found by calling .get(name of layer). This makes it easier to find a specific layer, and doesn't have to remember which number each layer has. The layer itself is in the same format as matrixMap.

Also added functionality for different sizes in maps. Previous iteration could only create map in a size of 5x5.